+++ lib/src/animated_fitted_text_field_container.dart
@@ -26,6 +26,18 @@ class AnimatedFittedTextFieldContainer extends StatefulWidget {
   /// `Curve` to use for animating shrinkage
   final Curve shrinkCurve;
 
+  /// The width of the `TextField.decoration.prefixIcon` if used
+  final double prefixIconWidth;
+
+  /// The width of the `TextField.decoration.suffixIcon` if used
+  final double suffixIconWidth;
+
+  /// The minimum width, if not set, the minimum width is 0 - i.e. there is no mimimum
+  final double minWidth;
+
+  /// The maximum width, if not set, the minimum width is infinity - i.e. there is no maximum
+  final double maxWidth;
+
   const AnimatedFittedTextFieldContainer({
     Key key,
     this.child,
@@ -33,6 +45,10 @@ class AnimatedFittedTextFieldContainer extends StatefulWidget {
     this.shrinkDuration = const Duration(milliseconds: 600),
     this.growCurve = Curves.easeOutCirc,
     this.shrinkCurve = Curves.easeInCirc,
+    this.prefixIconWidth = 48,
+    this.suffixIconWidth = 48,
+    this.minWidth,
+    this.maxWidth,
   }) : super(key: key);
   @override
   _AnimatedFittedTextFieldContainerState createState() =>
@@ -48,7 +64,7 @@ class _AnimatedFittedTextFieldContainerState
   @override
   void initState() {
     assert(widget.child.controller != null);
-    _textFieldWidth = calculateMaterialTextfieldWidth(widget.child);
+    _textFieldWidth = _geTextFieldWidth();
     widget.child.controller.addListener(_onTextChanged);
     _duration = widget.growDuration;
     _curve = widget.growCurve;
@@ -61,8 +77,22 @@ class _AnimatedFittedTextFieldContainerState
     super.dispose();
   }
 
+
+  double _geTextFieldWidth() {
+    double width = calculateMaterialTextfieldWidth(widget.child,
+        prefixIconWidth: widget.prefixIconWidth,
+        suffixIconWidth: widget.suffixIconWidth);
+    if (widget.minWidth != null && width < widget.minWidth) {
+      width = widget.minWidth;
+    }
+    if (widget.maxWidth != null && width > widget.maxWidth) {
+      width = widget.maxWidth;
+    }
+    return width;
+  }
+
   void _onTextChanged() {
-    double width = calculateMaterialTextfieldWidth(widget.child);
+    final width = _geTextFieldWidth();
     if (width != _textFieldWidth) {
       setState(() {
         if (width > _textFieldWidth) {