+++ lib/src/fitted_text_field_container.dart
@@ -14,8 +14,26 @@ class FittedTextFieldContainer extends StatefulWidget {
   /// The `TextField` to wrap
   final TextField child;
 
-  const FittedTextFieldContainer({Key key, @required this.child})
-      : assert(child != null),
+  /// The width of the `TextField.decoration.prefixIcon` if used
+  final double prefixIconWidth;
+
+  /// The width of the `TextField.decoration.suffixIcon` if used
+  final double suffixIconWidth;
+
+  /// The minimum width, if not set, the minimum width is 0 - i.e. there is no mimimum
+  final double minWidth;
+
+  /// The maximum width, if not set, the minimum width is infinity - i.e. there is no maximum
+  final double maxWidth;
+
+  const FittedTextFieldContainer({
+    Key key,
+    @required this.child,
+    this.prefixIconWidth = 48,
+    this.suffixIconWidth = 48,
+    this.minWidth,
+    this.maxWidth,
+  })  : assert(child != null),
         super(key: key);
   @override
   _FittedTextFieldContainerState createState() =>
@@ -28,7 +46,7 @@ class _FittedTextFieldContainerState extends State<FittedTextFieldContainer> {
   @override
   void initState() {
     assert(widget.child.controller != null);
-    _textFieldWidth = calculateMaterialTextfieldWidth(widget.child);
+    _textFieldWidth = _geTextFieldWidth();
     widget.child.controller.addListener(_onTextChanged);
     super.initState();
   }
@@ -39,8 +57,21 @@ class _FittedTextFieldContainerState extends State<FittedTextFieldContainer> {
     super.dispose();
   }
 
+  double _geTextFieldWidth() {
+    double width = calculateMaterialTextfieldWidth(widget.child,
+        prefixIconWidth: widget.prefixIconWidth,
+        suffixIconWidth: widget.suffixIconWidth);
+    if (widget.minWidth != null && width < widget.minWidth) {
+      width = widget.minWidth;
+    }
+    if (widget.maxWidth != null && width > widget.maxWidth) {
+      width = widget.maxWidth;
+    }
+    return width;
+  }
+
   void _onTextChanged() {
-    final double width = calculateMaterialTextfieldWidth(widget.child);
+    final width = _geTextFieldWidth();
     if (width != _textFieldWidth) {
       setState(() {
         _textFieldWidth = width;