+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 1.1.0
+* Added `prefixIconWidth` for specifying the width of the prefix icon (if used)
+* Added `suffixIconWidth` for specifying the width of the suffix icon (if used)
+* Added `minWidth` property for specifying a minimum width
+* Added `maxWidth` property for specifying a maximum width
+* Minimum width takes `labelText` and `hintText` into account (if specified) 
+
+## 1.0.0+3
+* Updated demo & documentation
+
 ## 1.0.0+2
 * Formatted code